renaming ugly.h to constants.h and removing libevalresp.c
ugly.h is not a friendly name therefore renamed it to constants.h to
better reflect what is in it. also wrapped each define in a ifndef so if
the user wants to change the the lengths at compile time it is possible

libevalresp.c has been rendered useless and therefore was removed.